+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support inline methods expanded via property invocations in dependency analysis.
 - Support for `with` statements.
+- Support for `System.MaxInt` compiler constant.
+- Support for `System.MaxLongInt` compiler constant.
+- Support for `System.CompilerVersion` compiler constant.
 
 ### Changed